requirements.txt
Flask==2.2.3
python-dotenv==1.0.1
gunicorn==20.1.0
Werkzeug==2.2.2
python-dateutil==2.8.2
requests==2.29.0
sqlitecloud==0.0.83
Frozen-Flask==1.0.2
formsg-python-sdk==0.1.13